services/tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Salim Ingram <salim.ingram@gmail.com>"]
readme = "README.md"
packages = [{include = "tasks_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"
mangum = "^0.17.0"
boto3 = "^1.26.150"
pyjwt = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
bandit = "^1.7.5"
moto = "^4.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
